+++ app/src/main/res/navigation/nav_graph.xml
@@ -22,7 +22,7 @@
         tools:layout="@layout/fragment_two">
         <argument
             android:name="repoInfo"
-            app:argType="jp.co.yumemi.android.code_check.data.source.network.RepoInfo" />
+            app:argType="jp.co.yumemi.android.code_check.domain.model.RepoInfo" />
     </fragment>
 
 </navigation>
